BACKEND/requirements.txt
# Core
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0

# API Documentation
drf-spectacular>=0.27

# Database drivers — install the one matching your DB_ENGINE:
# MySQL (XAMPP):   pip install mysqlclient
# PostgreSQL:      pip install psycopg2-binary
# SQL Server:      pip install mssql-django pyodbc
